requirements.txt
selenium
google-cloud-bigquery
pandas
webdriver-manager